NetBSD and gzip support
Includes two bits:

* Updates to the Makefile to aid in compiling under NetBSD
* Adds GZIP support to xfile, to allow reading/writing
  gzipped files directly.

It's a bit rough, and I'm not quite sure this is how I want the final
bits to be, but it works well enough to allow afsdump_scan to read an
entire 35GB gzipped volume dump.

Oh, and now includes the part to register GZIP as an xfile type.