Make afsdump_mtpt work for rw mountpoints
We should only be looking at the link target starting at the second
byte when matching on the cell name, since the first character can be
either a '#' or a '%'. We've already verified by this point that the
first char is either '#' or '%', so we can just ignore looking at it.