Make *VNodeData use 64-bit sizes
Currently, the CopyVNodeData and DumpVNodeData functions only use
32-bit file sizes, and only use the VTAG_DATA tag. As a result, if our
caller wants to generate a file greater than 4G, we silently only
print out some of the file data (a maximum of 2^32 bytes), and can
corrupt the output stream.

Fix these functions to always work with 64-bit sizes. For most
callers, the only change is that we need to pass in a size as &v->size
instead of v->size.