Partial revert "Allow compilation with newer OpenAFS"
This reverts much of commit 93307f1b5c898dbdb95184d72889d37b7bd0121c,
which addressed a com_err compatibility issue in a different way than
I already had.

It keeps various build fixes to afsdump_xsed, not related to com_err and
needed to otherwise bring it up to date.

Build system changes are also not reverted; they'll be handled in the
course of resolving the merge.